Election for Foobarnia
Need a way to tally results and determine winner or determine candidates for a run off election.

This commit includes a function to tally the results. The function checks for a majority winner as it counts each vote. If it has counted all the votes but has not found a majority then it takes the highest three candidates to enter them into a runoff. If these two conditions are not met then it will continue to count the vote as normal. A vote is counted by placing it into a voteObject. This allows the results to be sorted and keep the identity of the vote and its count together. As it is an object with two properties either property can be accessed and returned. After each vote is counted the results are sorted highest to lowest. This allows the condition for majority to be checked.