Mileage
Several errors preventing cars to be searched correctly. Desired ability to search by price range and fuel type.

Code for searching by mileage debugged by removing "nil" from the calculation using .compact. Not all cars in the database had a mileage defined therefore the method raised an error. Wrote a test to ensure that nil could be handled.
Defined a method to search by price range which would take in a min and max.
Included fuel in the attr_reader so it could then be used in the search methods.
Tests written for both new search methods.