Third test for third place tie passing
Created a while loop that starts at index position 3 and compares the count at each position with the 3rd place count. If the counts are equal the index adds one and it checks again. At the point where the counts do not match is where the splice function occurs. This is the case of no tie will splice at index 3 which returns the top three hence the second test will still pass.

Added a test to check whether if all candidates are tied that they all get returned for a run off.
This initially failed as the index tried to increase beyond the length of the result array. Therefore added a condition to the code to check that the index is less than the result.length.

All four tests are passing.